fix: Align gap_epoch types and add sentinel skip logic from master
- Change SequenceTracker::gap_epoch from uint16_t to uint32_t
- Change OrderState::gap_suppression_epoch from uint16_t to uint32_t
- Add sentinel skip logic: skip 0 on wrap-around since 0 is reserved
- This resolves the type conflict between master and FX-7054 branch

+++ src/core/order_state.hpp
@@ -75,7 +75,8 @@ struct OrderState {
     MismatchMask current_mismatch{};      // 1-byte mask
 
     std::uint32_t timer_generation{0};    // generation-based lazy cancel
-    std::uint16_t gap_suppression_epoch{0};  // Upgraded to uint16_t to avoid wrap-around issues
+    std::uint32_t gap_suppression_epoch{0};  // Using uint32_t to match SequenceTracker::gap_epoch
+                                             // IMPORTANT: Value 0 is reserved as sentinel (means "not flagged")
 
     // ===== FX-7054: Per-order gap uncertainty flags =====
     // Bitmask indicating which session gaps affect this order's reconciliation
